requirement.txt
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.1
cycler==0.11.0
fonttools==4.36.0
idna==3.3
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
wincertstore==0.2
